Separated out infer functionality into it's own service
